fix overflow. i32 -> i64
