+++ assembly/utils/index.ts
@@ -3,6 +3,6 @@ export * from './MovingAverages';
 export * from './Ranges';
 export * from './SlidingWindow';
 export * from './console';
-export * from './env';
+export * from './ccxt';
 export * from './types';
 export * from './triggers'
