+++ assembly/utils/types/Position.ts
@@ -1,11 +1,10 @@
 import { JSON } from 'json-as/assembly';
 
 @serializable
-export class Position 
-{
+export class Position {
     constructor(
-        public startTick : i32,
-        public endTick : i32,
-        public weight : i32)
-        {}
+        public startTick: i32,
+        public endTick: i32,
+        public weight: i32
+    ) { }
 }
